Split up integration tests.
