+++ test/endtoend/i18n-drag-drop/index.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>Uppy test page</title>
+  </head>
+  <body>
+    <style>
+      main {
+        max-width: 700px;
+        margin: auto;
+      }
+
+      #uppyi18n-progress {
+        position: relative;
+      }
+
+    </style>
+    <main>
+      <h2>Uppy DragDrop i18n + XHRUpload</h2>
+      <div>
+        <div id="uppyi18n"></div>
+        <div id="uppyi18n-progress"></div>
+      </div>
+    </main>
+
+    <link href="uppy.min.css" rel="stylesheet">
+    <script src="bundle.js"></script>
+  </body>
+</html>
