+++ test/endtoend/tus-drag-drop/index.html
@@ -12,33 +12,15 @@
         margin: auto;
       }
 
-      #uppyDragDrop-progress,
-      #uppyi18n-progress {
+      #uppyDragDrop-progress {
         position: relative;
       }
-
     </style>
     <main>
       <h2>Uppy DragDrop + Tus</h2>
       <div>
         <div id="uppyDragDrop"></div>
         <div id="uppyDragDrop-progress"></div>
-      </div>
-
-      <h2>Uppy DragDrop i18n + XHRUpload</h2>
-      <div>
-        <div id="uppyi18n"></div>
-        <div id="uppyi18n-progress"></div>
-      </div>
-
-      <h2>Uppy Dashboard + Tus</h2>
-      <div>
-        <div id="uppyDashboard"></div>
-      </div>
-
-      <h2>Uppy XHRUpload `limit`</h2>
-      <div>
-        <div id="uppyXhrLimit"></div>
       </div>
     </main>
