+++ test/endtoend/xhr-limit/test.js
@@ -3,7 +3,7 @@ var path = require('path')
 var http = require('http')
 var tempWrite = require('temp-write')
 
-var testURL = 'http://localhost:4567'
+var testURL = 'http://localhost:4567/xhr-limit'
 
 function uppySelectFakeFile (uppyID) {
   var blob = new Blob(
@@ -25,52 +25,7 @@ function browserSupportsChooseFile (capabilities) {
          capabilities.platformName !== 'Android'
 }
 
-describe('File upload with DragDrop + Tus, DragDrop + XHRUpload, i18n translated string', () => {
-  beforeEach(() => {
-    browser.url(testURL)
-  })
-
-  it('should upload a file with Tus and set progressbar to 100%', () => {
-    if (browserSupportsChooseFile(capabilities)) {
-      browser.chooseFile('#uppyDragDrop .uppy-DragDrop-input', path.join(__dirname, '../fixtures/image.jpg'))
-    } else {
-      browser.execute(uppySelectFakeFile, 'uppyDragDrop')
-    }
-    browser.pause(3000)
-    var html = browser.getHTML('#uppyDragDrop-progress .uppy-ProgressBar-percentage', false)
-    expect(parseInt(html)).to.be.equal(100)
-  })
-
-  it('should upload a file with XHRUpload and set progressbar to 100%', () => {
-    if (browserSupportsChooseFile(capabilities)) {
-      browser.chooseFile('#uppyi18n .uppy-DragDrop-input', path.join(__dirname, '../fixtures/image.jpg'))
-    } else {
-      browser.execute(uppySelectFakeFile, 'uppyi18n')
-    }
-    browser.pause(3000)
-    var html = browser.getHTML('#uppyi18n-progress .uppy-ProgressBar-percentage', false)
-    expect(parseInt(html)).to.be.equal(100)
-  })
-
-  it('should translate text strings into Russian', () => {
-    var text = browser.getText('#uppyi18n .uppy-DragDrop-label')
-    expect(text.trim()).to.be.equal('Перенесите файлы сюда или выберите')
-  })
-})
-
-  // it('another test', function () {
-  //   return browser
-  //     .url(uppyTestURL)
-  //     .chooseFile('#uppyDragDrop .uppy-DragDrop-input', path.join(__dirname, 'image.jpg'))
-  //     .pause(3000)
-  //     .getHTML('#uppyDragDrop-progress .UppyProgressBar-percentage', false).then(val => {
-  //       console.log(val)
-  //       expect(parseInt(val)).toBe(100)
-  //     })
-  // })
-// })
-
-describe.skip('XHRUpload with `limit`', () => {
+describe('XHRUpload with `limit`', () => {
   let server = null
   before(() => {
     server = http.createServer((req, res) => {
